[refactor] 기존 테스트 코드를 투표 생성 controller 단위 테스트, api 통합 테스트 코드로 분리 (#181)
